moved shaderUtils file to shaders folder, fixed width / height update
